refactor: extract loadSignerOrFail to utils/helius.ts
Move the session-then-disk keypair loading pattern into a shared helper
in utils/helius.ts. Replaces 3 duplicated copies in auth.ts and the
local copy in transfers.ts with a single reusable function.
